requirements.txt
tuyapy==0.1.4
fastapi>=0.115.4,<0.116.0
uvicorn>=0.32.0,<0.33.0
python-dotenv>=1.0.0
pydantic>=2.7.4,<3.0.0
requests>=2.31.0
python-multipart>=0.0.6
cryptography>=41.0.5
git+https://github.com/modelcontextprotocol/python-sdk.git
httpx==0.27.0
psutil>=5.9.0 
google-genai
chromadb>=0.5.0
sentence-transformers>=2.2.0
